dns & fixes
