+++ py-net-probe/deb/pre.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 #
-# Time-stamp: <2017-01-29 13:58:11 alex>
+# Time-stamp: <2017-05-08 18:06:41 alex>
 #
 # --------------------------------------------------------------------
 # PiProbe
@@ -21,3 +21,7 @@
 # --------------------------------------------------------------------
 
 rm -rf /home/pi/py-net-probe/*
+
+# v1.8.1
+# install dnspython
+pip install dnspython
