+++ py-net-probe/init.cfg
@@ -1,6 +1,6 @@
 # configuration file for the probe client
 #
-# Time-stamp: <2017-04-29 17:00:46 alex>
+# Time-stamp: <2017-05-08 17:57:22 alex>
 #
 # the scheduler section is loaded by the piprobe
 
@@ -9,8 +9,8 @@ get_conf: 3600
 push_results: 15
 ping_server: 10
 check_probes: 30
-stats_probes: 60
-stats_push: 300
+stats_probes: 5
+stats_push: 10
 upgrade: 21600
 
 [dev_scheduler]
