+++ py-net-probe/probelib/probemain.py
@@ -1,6 +1,6 @@
 # -*- Mode: Python; python-indent-offset: 4 -*-
 #
-# Time-stamp: <2017-04-30 17:28:25 alex>
+# Time-stamp: <2017-05-08 18:22:36 alex>
 #
 # --------------------------------------------------------------------
 # PiProbe
@@ -220,3 +220,21 @@ class probemain(object):
 
         """
         return True
+
+    # -----------------------------------------
+    @classmethod
+    def f_testv4(cls, data):
+        """testing method for insertion in the job list, check if ip version 4
+
+        """
+        return data['version'] == 4
+
+    # -----------------------------------------
+    @classmethod
+    def _setValueFromConfig(cls, _config, field, default=0, _min=-1000, _max=1000):
+        if _config.__contains__(field):
+            r = _config[field]
+            r = max(_min, r)
+            r = min(_max, r)
+            return r
+        return default
